Install Compass in virtual env
Create virtualenv for compass, point compass site packages to virtualenv,
make virtualenv accessable for apache.
Change compass executable scripts to use python in env to be virtualenv
compatible.
Make progress_update daemon use virtualenv.
Comment out some health checks that use packages not available to pip, eg,
yum, apt since we have other ways to check packages any way.

Conflicts:

	conf/compassd
	install/compass.sh
	install/dependency.sh
	install/prepare.sh
	misc/apache/compass.wsgi
	requirements.txt

+++ install/prepare.sh
@@ -24,6 +24,12 @@ copy2dir()
             cd $destdir
             git remote set-url origin $repo
             git remote update
+            if [ $? -ne 0 ]; then
+                echo "failed to git remote update $repo in $destdir"
+                exit 1
+            else
+                echo "git remote update $repo in $destdir succeeded"
+            fi
             git reset --hard
             git clean -x -f
             git checkout $git_branch
@@ -45,11 +51,19 @@ copy2dir()
             if [[ ! -z $ZUUL_REF ]]; then
                 git_repo=$ZUUL_URL/$3
                 git_ref=$ZUUL_REF
-                git_branch=$ZUUL_BRANCH
+                if git branch -a|grep ${ZUUL_BRANCH}; then
+                    git_branch=$ZUUL_BRANCH
+                else
+                    git_branch=master
+                fi
             elif [[ ! -z $GERRIT_REFSPEC ]]; then
                 git_repo=https://$GERRIT_HOST/$3
                 git_ref=$GERRIT_REFSPEC
-                git_branch=$GERRIT_BRANCH
+                if git branch -a|grep $GERRIT_BRANCH; then
+                    git_branch=$GERRIT_BRANCH
+                else
+                    git_branch=master
+                fi
             fi
             git reset --hard remotes/origin/$git_branch
             git fetch $git_repo $git_ref && git checkout FETCH_HEAD
@@ -141,18 +155,53 @@ else
     echo "squid conf is updated"
 fi
 
+#update mysqld
+sudo service mysqld restart
+MYSQL_USER=${MYSQL_USER:-root}
+MYSQL_OLD_PASSWORD=${MYSQL_OLD_PASSWORD:-}
+MYSQL_PASSWORD=${MYSQL_PASSWORD:-root}
+MYSQL_SERVER=${MYSQL_SERVER:-127.0.0.1}
+MYSQL_PORT=${MYSQL_PORT:-3306}
+MYSQL_DATABASE=${MYSQL_DATABASE:-db}
+# first time set mysql password
+sudo mysqladmin -h${MYSQL_SERVER} --port=${MYSQL_PORT} -u ${MYSQL_USER} password ${MYSQL_PASSWORD}
+sudo mysqladmin -h${MYSQL_SERVER} --port=${MYSQL_PORT} -u ${MYSQL_USER} -p"${MYSQL_OLD_PASSWORD}" password ${MYSQL_PASSWORD}
+mysql -h${MYSQL_SERVER} --port=${MYSQL_PORT} -u${MYSQL_USER} -p${MYSQL_PASSWORD} -e "show databases;"
+if [[ "$?" != "0" ]]; then
+    echo "mysql password set failed"
+    exit 1
+else
+    echo "mysql password set successful"
+fi
+sudo mysql -h${MYSQL_SERVER} --port=${MYSQL_PORT} -u${MYSQL_USER} -p${MYSQL_PASSWORD} -e "drop database ${MYSQL_DATABASE}"
+sudo mysql -h${MYSQL_SERVER} --port=${MYSQL_PORT} -u${MYSQL_USER} -p${MYSQL_PASSWORD} -e "create database ${MYSQL_DATABASE}"
+if [[ "$?" != "0" ]]; then
+    echo "mysql database set fails"
+    exit 1
+else
+    echo "mysql database set succeeds"
+fi
+sudo service mysqld restart
+sudo service mysqld status
+if [[ "$?" != "0" ]]; then
+    echo "mysqld is not started"
+    exit 1
+else
+    echo "mysqld conf is updated" 
+fi
+
 cd $SCRIPT_DIR
 if [ -z $WEB_SOURCE ]; then
     echo "web source $WEB_SOURCE is not set"
     exit 1
 fi
-copy2dir "$WEB_SOURCE" "$WEB_HOME" "stackforge/compass-web"
+copy2dir "$WEB_SOURCE" "$WEB_HOME" "stackforge/compass-web" || exit $?
 
 if [ -z $ADAPTERS_SOURCE ]; then
     echo "adpaters source $ADAPTERS_SOURCE is not set"
     exit 1
 fi
-copy2dir "$ADAPTERS_SOURCE" "$ADAPTERS_HOME" "stackforge/compass-adapters"
+copy2dir "$ADAPTERS_SOURCE" "$ADAPTERS_HOME" "stackforge/compass-adapters" || exit $?
 
 if [ "$tempest" == "true" ]; then
     if [[ ! -e /tmp/tempest ]]; then
@@ -160,6 +209,8 @@ if [ "$tempest" == "true" ]; then
         if [[ "$?" != "0" ]]; then
             echo "failed to git clone tempest project"
             exit 1
+        else
+            echo "git clone tempest project succeeded"
         fi
         cd /tmp/tempest
         git checkout grizzly-eol
@@ -167,12 +218,54 @@ if [ "$tempest" == "true" ]; then
         cd /tmp/tempest
         git remote set-url origin http://git.openstack.org/openstack/tempest
         git remote update
+        if [[ "$?" != "0" ]]; then
+            echo "failed to git remote update tempest project"
+            exit 1
+        else
+            echo "git remote update tempest project succeeded"
+        fi
         git reset --hard
         git clean -x -f -d -q
         git checkout grizzly-eol
     fi
+    source `which virtualenvwrapper.sh`
+    if ! lsvirtualenv |grep tempest>/dev/null; then
+        mkvirtualenv tempest
+    fi
+    workon tempest
     cd /tmp/tempest
     pip install -e .
+    pip install sqlalchemy
+    if [[ "$?" != "0" ]]; then
+        echo "failed to install tempest project"
+        deactivate
+        exit 1
+    else
+        echo "install tempest project succeeded"
+        deactivate
+    fi
+fi
+
+source `which virtualenvwrapper.sh`
+if ! lsvirtualenv |grep compass-core>/dev/null; then
+    mkvirtualenv compass-core
+fi
+workon compass-core
+cd $COMPASSDIR
+pip install -U -r requirements.txt
+if [[ "$?" != "0" ]]; then
+    echo "failed to install compass requiremnts"
+    deactivate
+    exit 1
+fi
+pip install -U -r test-requirements.txt
+if [[ "$?" != "0" ]]; then
+    echo "failed to install compass test requiremnts"
+    deactivate
+    exit 1
+else
+    echo "intall compass requirements succeeded"
+    deactivate
 fi
 
 download()
@@ -185,13 +278,13 @@ download()
     else
         if [[ "$url" =~ (http|https|ftp):// ]]; then
             echo "downloading $url to /tmp/${package}"
-            wget -c --progress=bar:force -O /tmp/${package}.tmp $url
+            curl -L -o /tmp/${package}.tmp $url
             if [[ "$?" != "0" ]]; then
                 echo "failed to download $package"
                 exit 1
             else
                 echo "successfully download $package"
-                cp -rf /tmp/${package}.tmp /tmp/${package}
+                mv -f /tmp/${package}.tmp /tmp/${package}
             fi
         else
             cp -rf $url /tmp/${package}
@@ -239,7 +332,7 @@ wget-1.12-1.8.${CENTOS_IMAGE_TYPE_OTHER}${CENTOS_IMAGE_VERSION_MAJOR}.${CENTOS_I
 ntpdate-4.2.6p5-1.${CENTOS_IMAGE_TYPE_OTHER}${CENTOS_IMAGE_VERSION_MAJOR}.${CENTOS_IMAGE_TYPE,,}.${CENTOS_IMAGE_ARCH}.rpm"
 
 for f in $centos_ppa_repo_packages; do
-    download ftp://rpmfind.net/linux/${IMAGE_TYPE,,}/${IMAGE_VERSION_MAJOR}/os/${IMAGE_ARCH}/Packages/$f $f || exit $?
+    download http://rpmfind.net/linux/${IMAGE_TYPE,,}/${IMAGE_VERSION_MAJOR}/os/${IMAGE_ARCH}/Packages/$f $f || exit $?
 done
 
 centos_ppa_repo_rsyslog_packages="
@@ -265,16 +358,24 @@ download "$CENTOS_IMAGE_SOURCE" ${CENTOS_IMAGE_NAME}-${CENTOS_IMAGE_ARCH}.iso ||
 download "$UBUNTU_IMAGE_SOURCE" ${UBUNTU_IMAGE_NAME}-${UBUNTU_IMAGE_ARCH}.iso || exit $?
 
 # Install net-snmp
-sudo cp -rn /etc/snmp/snmp.conf /root/backup/
+if [[ ! -e /etc/snmp ]]; then
+    sudo mkdir -p /etc/snmp
+fi
+if [[ -e /etc/snmp/snmp.conf ]]; then
+    sudo cp -rn /etc/snmp/snmp.conf /root/backup/
+    sudo rm -f /etc/snmp/snmp.conf
+fi
 sudo mkdir -p /usr/local/share/snmp/
 sudo cp -rf $COMPASSDIR/mibs /usr/local/share/snmp/
-sudo rm -f /etc/snmp/snmp.conf
 sudo cp -rf $COMPASSDIR/misc/snmp/snmp.conf /etc/snmp/snmp.conf
 sudo chmod 644 /etc/snmp/snmp.conf
 sudo mkdir -p /var/lib/net-snmp/mib_indexes
 sudo chmod 755 /var/lib/net-snmp/mib_indexes
 
 # generate ssh key
+if [[ ! -e $HOME/.ssh ]]; then
+    sudo mkdir -p $HOME/.ssh
+fi
 if [ ! -e $HOME/.ssh/id_rsa.pub ]; then
     rm -rf $HOME/.ssh/id_rsa
     ssh-keygen -t rsa -f $HOME/.ssh/id_rsa -q -N ''
